+++ crates/fluvio-smartengine/src/engine/wasmtime_engine/memory.rs
