+++ crates/fluvio-smartengine/src/engine/wasmtime_engine/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod memory;
+pub(crate) mod transforms;
+pub(crate) mod init;
+pub(crate) mod state;
+pub(crate) mod engine;
+pub(crate) mod instance;
+pub use engine::{SmartEngineImp, initialize_imp, SmartModuleChainInstanceImp};
+
+use super::*;
