+++ crates/fluvio-smartengine/src/engine/wasmtime_engine/transforms/array_map.rs
@@ -10,21 +10,22 @@ mod test {
 
     use crate::engine::{
         SmartEngine, SmartModuleChainBuilder, SmartModuleConfig, metrics::SmartModuleChainMetrics,
-        transforms::simple_transform::MAP_FN_NAME,
+        wasmtime_engine::transforms::simple_transform::ARRAY_MAP_FN_NAME,
     };
-    use crate::engine::fixture::read_wasm_module;
 
-    const SM_MAP: &str = "fluvio_smartmodule_map";
+    const SM_ARRAY_MAP: &str = "fluvio_smartmodule_array_map_array";
+
+    use crate::engine::fixture::read_wasm_module;
 
     #[ignore]
     #[test]
-    fn test_map() {
+    fn test_array_map() {
         let engine = SmartEngine::new();
         let mut chain_builder = SmartModuleChainBuilder::default();
 
         chain_builder.add_smart_module(
             SmartModuleConfig::builder().build().unwrap(),
-            read_wasm_module(SM_MAP),
+            read_wasm_module(SM_ARRAY_MAP),
         );
 
         let mut chain = chain_builder
@@ -34,16 +35,18 @@ mod test {
 
         assert_eq!(
             chain.instances().first().expect("first").transform().name(),
-            MAP_FN_NAME
+            ARRAY_MAP_FN_NAME
         );
 
         let metrics = SmartModuleChainMetrics::default();
-        let input = vec![Record::new("apple"), Record::new("fruit")];
+
+        let input = vec![Record::new("[\"Apple\",\"Banana\",\"Cranberry\"]")];
         let output = chain
             .process(SmartModuleInput::try_from(input).expect("input"), &metrics)
             .expect("process");
-        assert_eq!(output.successes.len(), 2); // one record passed
-        assert_eq!(output.successes[0].value.as_ref(), b"APPLE");
-        assert_eq!(output.successes[1].value.as_ref(), b"FRUIT");
+        assert_eq!(output.successes.len(), 3); // generate 3 records
+        assert_eq!(output.successes[0].value.as_ref(), b"\"Apple\"");
+        assert_eq!(output.successes[1].value.as_ref(), b"\"Banana\"");
+        assert_eq!(output.successes[2].value.as_ref(), b"\"Cranberry\"");
     }
 }
